refactor: Criando o TransacaoMapper
Essa classe tem como função prover o metodo toResponse para o transacaoService, primeira função delegada do TransacaoService
